refactor: Refactor ingestion to only have one ingestion task
